Factor out common html library code
