+++ lib/ui/cardano-wallet-ui.cabal
@@ -33,19 +33,19 @@ flag release
 library
   import:           language, opts-lib
   exposed-modules:
+    Cardano.Wallet.UI.Common.Html.Html
+    Cardano.Wallet.UI.Common.Html.Htmx
+    Cardano.Wallet.UI.Common.Html.Lib
     Cardano.Wallet.UI.Cookies
     Cardano.Wallet.UI.Lib.ListOf
     Cardano.Wallet.UI.Personal.API
     Cardano.Wallet.UI.Personal.Handlers.Addresses
     Cardano.Wallet.UI.Personal.Handlers.Lib
     Cardano.Wallet.UI.Personal.Handlers.Settings
-    Cardano.Wallet.UI.Personal.Handlers.SSE
+    Cardano.Wallet.UI.Common.Handlers.SSE
     Cardano.Wallet.UI.Personal.Handlers.State
     Cardano.Wallet.UI.Personal.Handlers.Wallet
     Cardano.Wallet.UI.Personal.Handlers.Wallets
-    Cardano.Wallet.UI.Personal.Html.Html
-    Cardano.Wallet.UI.Personal.Html.Htmx
-    Cardano.Wallet.UI.Personal.Html.Lib
     Cardano.Wallet.UI.Personal.Html.Pages.About
     Cardano.Wallet.UI.Personal.Html.Pages.Addresses
     Cardano.Wallet.UI.Personal.Html.Pages.Lib
