+++ lib/ui/src/Cardano/Wallet/UI/Personal/Handlers/Addresses.hs
@@ -36,12 +36,12 @@ import Cardano.Wallet.Api.Types
 import Cardano.Wallet.Primitive.NetworkId
     ( HasSNetworkId (..)
     )
+import Cardano.Wallet.UI.Common.Html.Html
+    ( RawHtml (..)
+    )
 import Cardano.Wallet.UI.Personal.Handlers.Lib
     ( alertOnServerError
     , catching
-    )
-import Cardano.Wallet.UI.Personal.Html.Html
-    ( RawHtml (..)
     )
 import Cardano.Wallet.UI.Personal.Layer
     ( SessionLayer (..)
