+++ lib/ui/src/Cardano/Wallet/UI/Personal/Server.hs
@@ -49,6 +49,10 @@ import Cardano.Wallet.Primitive.NetworkId
 import Cardano.Wallet.Shelley.BlockchainSource
     ( BlockchainSource (..)
     )
+import Cardano.Wallet.UI.Common.Html.Html
+    ( RawHtml (..)
+    , renderHtml
+    )
 import Cardano.Wallet.UI.Cookies
     ( CookieResponse
     , RequestCookies
@@ -65,7 +69,7 @@ import Cardano.Wallet.UI.Personal.Handlers.Addresses
 import Cardano.Wallet.UI.Personal.Handlers.Settings
     ( toggleSSE
     )
-import Cardano.Wallet.UI.Personal.Handlers.SSE
+import Cardano.Wallet.UI.Common.Handlers.SSE
     ( sse
     )
 import Cardano.Wallet.UI.Personal.Handlers.State
@@ -80,10 +84,6 @@ import Cardano.Wallet.UI.Personal.Handlers.Wallet
     )
 import Cardano.Wallet.UI.Personal.Handlers.Wallets
     ( listWallets
-    )
-import Cardano.Wallet.UI.Personal.Html.Html
-    ( RawHtml (..)
-    , renderHtml
     )
 import Cardano.Wallet.UI.Personal.Html.Pages.Addresses
     ( addressesH
